Add cost estimate to README
