+++ agents-for-bedrock/use-case-examples/insurance-claim-lifecycle-automation/shell/create-customer-resources.sh
@@ -15,7 +15,7 @@ export GATHER_EVIDENCE_KEY="agent/lambda/action-groups/gather_evidence.zip"
 export SEND_REMINDER_KEY="agent/lambda/action-groups/send_reminder.zip"
 
 aws s3 mb s3://${ARTIFACT_BUCKET_NAME} --region ${AWS_REGION}
-aws s3 cp ../agent/ s3://${ARTIFACT_BUCKET_NAME}/agent/ --region ${AWS_REGION} --recursive --exclude ".DS_Store"
+aws s3 cp ../agent/ s3://${ARTIFACT_BUCKET_NAME}/agent/ --region ${AWS_REGION} --recursive --exclude ".DS_Store" --exclude "*/.DS_Store"
 
 export BEDROCK_AGENTS_LAYER_ARN=$(aws lambda publish-layer-version \
     --layer-name bedrock-agents \
